refactor: command 객체 패키지 이동 및 적용
