+++ api/src/main/java/io/devarium/api/controller/feedback/dto/UpdateQuestionRequest.java
@@ -2,7 +2,7 @@ package io.devarium.api.controller.feedback.dto;
 
 import io.devarium.core.domain.feedback.question.QuestionContent;
 import io.devarium.core.domain.feedback.question.QuestionType;
-import io.devarium.core.domain.feedback.question.port.UpdateQuestion;
+import io.devarium.core.domain.feedback.question.command.UpdateQuestion;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
