+++ api/src/main/java/io/devarium/api/controller/project/dto/UpsertProjectRequest.java
@@ -1,6 +1,6 @@
 package io.devarium.api.controller.project.dto;
 
-import io.devarium.core.domain.project.port.UpsertProject;
+import io.devarium.core.domain.project.command.UpsertProject;
 import io.devarium.core.domain.skill.Skill;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
