+++ api/src/main/java/io/devarium/api/controller/reply/dto/UpsertReplyRequest.java
@@ -1,6 +1,6 @@
 package io.devarium.api.controller.reply.dto;
 
-import io.devarium.core.domain.reply.port.UpsertReply;
+import io.devarium.core.domain.reply.command.UpsertReply;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
