+++ api/src/main/java/io/devarium/api/controller/team/dto/UpdateLeaderRequest.java
@@ -1,6 +1,6 @@
 package io.devarium.api.controller.team.dto;
 
-import io.devarium.core.domain.team.port.UpdateLeader;
+import io.devarium.core.domain.team.command.UpdateLeader;
 import jakarta.validation.constraints.NotNull;
 
 public record UpdateLeaderRequest(
