+++ api/src/main/java/io/devarium/api/controller/user/dto/UpdateUserRequest.java
@@ -1,6 +1,6 @@
 package io.devarium.api.controller.user.dto;
 
-import io.devarium.core.domain.user.port.UpdateUser;
+import io.devarium.core.domain.user.command.UpdateUser;
 
 public record UpdateUserRequest(
     String username,
