+++ core/src/main/java/io/devarium/core/domain/comment/command/UpsertComment.java
@@ -1,4 +1,4 @@
-package io.devarium.core.domain.comment.port;
+package io.devarium.core.domain.comment.command;
 
 public interface UpsertComment {
