+++ core/src/main/java/io/devarium/core/domain/feedback/question/command/CreateQuestion.java
@@ -1,4 +1,4 @@
-package io.devarium.core.domain.feedback.question.port;
+package io.devarium.core.domain.feedback.question.command;
 
 import io.devarium.core.domain.feedback.question.QuestionContent;
 import io.devarium.core.domain.feedback.question.QuestionType;
