+++ core/src/main/java/io/devarium/core/domain/feedback/question/command/UpdateQuestionOrders.java
@@ -1,4 +1,4 @@
-package io.devarium.core.domain.feedback.question.port;
+package io.devarium.core.domain.feedback.question.command;
 
 import java.util.List;
