+++ core/src/main/java/io/devarium/core/domain/member/command/CreateMembers.java
@@ -1,4 +1,4 @@
-package io.devarium.core.domain.member.port;
+package io.devarium.core.domain.member.command;
 
 import java.util.Set;
