+++ core/src/main/java/io/devarium/core/domain/reply/command/UpsertReply.java
@@ -1,4 +1,4 @@
-package io.devarium.core.domain.reply.port;
+package io.devarium.core.domain.reply.command;
 
 public interface UpsertReply {
